Fix authentication
What's up
Part of the effort in https://github.com/alphasights/lion/pull/10 was to update the authentication scheme. This new scheme relied on [Torii](http://vestorly.github.io/torii/) for its out-of-the-box github oauth2 provider implementation, as well as [Ember Simple Auth](http://ember-simple-auth.com/) to authenticate and authorize requests to the server.

However, the latest version on Ember Simple Auth came with a lot of breaking changes. These made updating the project as a whole very difficult. Luckily, http://log.simplabs.com/post/131698328145/updating-to-ember-simple-auth-10 explains how to migrate with reduced hair-pulling.

What this does
- Updates `ember-simple-auth` and its usages
- Using `ember-simple-auth`'s new conventions, gets the auth functioning properly.